# src/swaggerOutput.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "TradeMark",
    "description": "TradeMark REST API Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/",
      "description": "Local Host"
    }
  ],
  "paths": {
...
    "/user/profile": {
      "get": {
        "tags": ["User"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": ["User"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          }
        }
      }
    },
...
    "/transaction/all": {
      "get": {
        "tags": ["User Transactions"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/transaction/{id}": {
      "get": {
        "tags": ["User Transactions"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/plan/all": {
      "get": {
        "tags": ["Plans"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlansResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/PlansResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/plan/{id}": {
      "get": {
        "tags": ["Plans"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/admin/plans/{id}": {
      "delete": {
        "tags": ["Admin - Plans"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlansResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/PlansResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "406": {
            "description": "Not Acceptable"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProfileRequest": {
        "type": "object",
        "properties": {
          "bitcoin": {
            "type": "string",
            "example": "string"
          },
          "sol": {
            "type": "string",
            "example": "string"
          },
          "ethereum": {
            "type": "string",
            "example": "string"
          },
          "profilePic": {
            "type": "string",
            "example": "string"
          },
          "phoneNumber": {
            "type": "string",
            "example": "string"
          },
          "dob": {
            "type": "string",
            "example": "string"
          },
          "state": {
            "type": "string",
            "example": "string"
          },
          "city": {
            "type": "string",
            "example": "string"
          },
          "bankName": {
            "type": "string",
            "example": "string"
          },
          "accountNumber": {
            "type": "string",
            "example": "string"
          },
          "routingNumber": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "bitcoin",
          "sol",
          "ethereum",
          "profilePic",
          "phoneNumber",
          "dob",
          "state",
          "city",
          "bankName",
          "accountNumber",
          "routingNumber"
        ],
        "xml": {
          "name": "ProfileRequest"
        }
      },
...
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "string"
              },
              "name": {
                "type": "string",
                "example": "string"
              },
              "username": {
                "type": "string",
                "example": "string"
              },
              "uid": {
                "type": "string",
                "example": "string"
              },
              "email": {
                "type": "string",
                "example": "string@m.m"
              },
              "password": {
                "type": "string",
                "example": "string"
              },
              "balance": {
                "type": "string",
                "example": "number"
              },
              "activeInvestments": {
                "type": "string",
                "example": "number"
              },
              "pendingWithdrawals": {
                "type": "string",
                "example": "number"
              },
              "earnings": {
                "type": "string",
                "example": "number"
              },
              "verified": {
                "type": "string",
                "example": "boolean"
              },
              "referralId": {
                "type": "string",
                "example": "string"
              },
              "bitcoin": {
                "type": "string",
                "example": "string"
              },
              "sol": {
                "type": "string",
                "example": "string"
              },
              "ethereum": {
                "type": "string",
                "example": "string"
              },
              "type": {
                "type": "string",
                "example": "admin | user"
              },
              "createdAt": {
                "type": "string",
                "example": "number"
              },
              "updatedAt": {
                "type": "string",
                "example": "number"
              }
            }
          }
        },
        "xml": {
          "name": "ProfileResponse"
        }
      },
...
      "TransactionsResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "string"
                },
                "uid": {
                  "type": "string",
                  "example": "string"
                },
                "amount": {
                  "type": "string",
                  "example": "number"
                },
                "from": {
                  "type": "string",
                  "example": "string"
                },
                "to": {
                  "type": "string",
                  "example": "string"
                },
                "plan_name": {
                  "type": "string",
                  "example": "string"
                },
                "plan_id": {
                  "type": "string",
                  "example": "string"
                },
                "investment_id": {
                  "type": "string",
                  "example": "string"
                },
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "status": {
                  "type": "string",
                  "example": "string"
                },
                "createdAt": {
                  "type": "string",
                  "example": "number"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "number"
                }
              }
            }
          }
        },
        "xml": {
          "name": "TransactionsResponse"
        }
      },
      "TransactionResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "string"
              },
              "uid": {
                "type": "string",
                "example": "string"
              },
              "amount": {
                "type": "string",
                "example": "number"
              },
              "from": {
                "type": "string",
                "example": "string"
              },
              "to": {
                "type": "string",
                "example": "string"
              },
              "plan_name": {
                "type": "string",
                "example": "string"
              },
              "plan_id": {
                "type": "string",
                "example": "string"
              },
              "investment_id": {
                "type": "string",
                "example": "string"
              },
              "type": {
                "type": "string",
                "example": "string"
              },
              "status": {
                "type": "string",
                "example": "string"
              },
              "createdAt": {
                "type": "string",
                "example": "number"
              },
              "updatedAt": {
                "type": "string",
                "example": "number"
              }
            }
          }
        },
        "xml": {
          "name": "TransactionResponse"
        }
      },
      "PlansResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "returns": {
                  "type": "string",
                  "example": "string"
                },
                "minAmount": {
                  "type": "string",
                  "example": "string"
                },
                "duration": {
                  "type": "string",
                  "example": "string"
                },
                "maxAmount": {
                  "type": "string",
                  "example": "string"
                },
                "uid": {
                  "type": "string",
                  "example": "string"
                }
              },
              "required": [
                "name",
                "returns",
                "minAmount",
                "duration",
                "maxAmount",
                "uid"
              ]
            }
          }
        },
        "xml": {
          "name": "PlansResponse"
        }
      },
      "PlanResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "string"
              },
              "returns": {
                "type": "string",
                "example": "string"
              },
              "minAmount": {
                "type": "string",
                "example": "string"
              },
              "duration": {
                "type": "string",
                "example": "string"
              },
              "maxAmount": {
                "type": "string",
                "example": "string"
              },
              "uid": {
                "type": "string",
                "example": "string"
              }
            },
            "required": [
              "name",
              "returns",
              "minAmount",
              "duration",
              "maxAmount",
              "uid"
            ]
          }
        },
        "xml": {
          "name": "PlanResponse"
        }
      },
...
      "InvalidToken": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid Token."
          }
        },
        "xml": {
          "name": "InvalidToken"
        }
      },
...
      "NotExists": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Does not exists."
          }
        },
        "xml": {
          "name": "NotExists"
        }
      },
      "InvalidID": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid ID."
          }
        },
        "xml": {
          "name": "InvalidID"
        }
      },
...
    }
  },
...
}